% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as_featurecollection.R
\name{as_featurecollection}
\alias{as_featurecollection}
\title{As featurecollection}
\usage{
as_featurecollection(x)
}
\arguments{
\item{x}{Input}
}
\description{
As featurecollection
}
\details{
Helper function to make a FeatureCollection list object for use
in vizualizing, e.g., with \code{\link{leaflet}}
}
\examples{
str <- 'MULTIPOINT ((100.000 3.101), (101.000 2.100), (3.140 2.180),
(31.140 6.180), (31.140 78.180))'
wkt2geojson(str, fmt = 2) \%>\%
  as_featurecollection() \%>\%
  lawn::view()
}

