% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wkt_r.R
\name{dump_linestring}
\alias{dump_linestring}
\title{Convert GeoJSON-like LineString object to WKT.}
\usage{
dump_linestring(obj, fmt = 16)
}
\arguments{
\item{obj}{A GeoJSON-like `dict` representing a LineString}

\item{fmt}{Format string which indicates the number of digits to display after the
decimal point when formatting coordinates.}
}
\description{
Convert GeoJSON-like LineString object to WKT.
}
\examples{
st <- list(type = 'LineString',
           coordinates = list(c(0.0, 0.0, 10.0), c(2.0, 1.0, 20.0),
                             c(4.0, 2.0, 30.0), c(5.0, 4.0, 40.0)))
dump_linestring(st, fmt=0)
}

