% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lint.R
\name{lint}
\alias{lint}
\title{Validate WKT strings}
\usage{
lint(str)
}
\arguments{
\item{str}{A WKT string}
}
\value{
A logical (\code{TRUE} or \code{FALSE})
}
\description{
Validate WKT strings
}
\examples{
lint("POINT (1 2)")
lint("POINT (1 2 3)")
lint("LINESTRING EMPTY")
lint("LINESTRING (100 0, 101 1)")
}

