% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wkt2geojson.R
\name{load_multipolygon}
\alias{load_multipolygon}
\title{Convert WKT to GeoJSON-like MULTIPOLYGON object.}
\usage{
load_multipolygon(str, fmt = 16, feature = TRUE)
}
\arguments{
\item{str}{A GeoJSON-like object representing a Point, LineString, Polygon, MultiPolygon, etc.}

\item{fmt}{Format string which indicates the number of digits to display after the
decimal point when formatting coordinates.}

\item{feature}{(logical) Make a feature geojson object. Default: TRUE}
}
\description{
Convert WKT to GeoJSON-like MULTIPOLYGON object.
}
\keyword{internal}

