% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/polygon.R
\name{polygon}
\alias{polygon}
\title{Make WKT polygon objects}
\usage{
polygon(..., fmt = 16)
}
\arguments{
\item{...}{A GeoJSON-like object representing a Point, LineString, Polygon, MultiPolygon, etc.}

\item{fmt}{Format string which indicates the number of digits to display after the
decimal point when formatting coordinates. Max: 20}
}
\description{
Make WKT polygon objects
}
\examples{
## empty polygon
polygon("empty")
# polygon("stuff")

# numeric
polygon(c(100.001, 0.001), c(101.12345, 0.001), c(101.001, 1.001), c(100.001, 0.001), fmt=2)

# data.frame
df <- us_cities[2:5,c('long','lat')]
df <- rbind(df, df[1,])
polygon(df) \%>\% wktview(zoom=4)

# list
polygon(list(c(100.001, 0.001), c(101.12345, 0.001), c(101.001, 1.001), c(100.001, 0.001))) \%>\%
  wktview(zoom=7)
}

