% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_convert.R
\name{sf_convert}
\alias{sf_convert}
\title{Convert spatial objects to WKT}
\usage{
sf_convert(x)
}
\arguments{
\item{x}{an sf or sfc object. one or more can be submitted}
}
\value{
a character vector of WKT objects
}
\description{
\code{sp_convert} turns objects from the \code{sp} package
(SpatialPolygons, SpatialPolygonDataFrames) or the \code{sf} package
(sf, sfc, POLYGON, MULTIPOLYGON) - into WKT POLYGONs or MULTIPOLYGONs
}
