% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_convert.R
\name{sp_convert}
\alias{sp_convert}
\alias{sf_convert}
\title{Convert spatial objects to WKT}
\usage{
sp_convert(x, group = TRUE)

sf_convert(x, group = TRUE)
}
\arguments{
\item{x}{for \code{sp_convert()}, a list of SP/SPDF objects (or a single object)
for \code{sf_convert()}, an sf, sfc, POLYGON, or MULTIPOLYGON sf object}

\item{group}{whether or not to group coordinates together in the case
that an object in \code{x} has multiple sets of coordinates. If \code{TRUE}
(the default), such objects will be returned as \code{MULTIPOLYGON}'s
\itemize{
\item if \code{FALSE}, as a vector of \code{POLYGON}'s
}}
}
\value{
either a character vector of WKT objects - one per sp object -
if \code{group} is \code{TRUE}, or a list of vectors if \code{group} is \code{FALSE}
}
\description{
\code{sp_convert} turns objects from the \code{sp} package
(SpatialPolygons, SpatialPolygonDataFrames) or the \code{sf} package
(sf, sfc, POLYGON, MULTIPOLYGON) - into WKT POLYGONs or MULTIPOLYGONs
}
\seealso{
\code{\link[=bounding_wkt]{bounding_wkt()}}, for turning bounding boxes within
\code{sp} objects into WKT objects.
}
