% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vis.R
\name{vis}
\alias{vis}
\title{Visualize well-known text area's on a map using leaflet and htmlwidgets packages}
\description{
Visualize well-known text area's on a map using leaflet and htmlwidgets packages
}
\examples{
\dontrun{
# Using htmlwidgets and leaflet packages
library("leaflet")
library("htmlwidgets")

## points
str <- "POINT (-116.4000000000000057 45.2000000000000028)"
geojson <- wkt2geojson(str) \%>\% properties(style=list(color = "red"))
leaflet() \%>\%
 addTiles() \%>\%
 addGeoJSON(geojson)

## polygons
geojson <- wkt2geojson(str) \%>\% properties(style=list(color = "red"))
cen <- centroid(geojson)
leaflet() \%>\%
 addTiles() \%>\%
 setView(lng = cen[2], lat = cen[1], zoom = 6) \%>\%
 addGeoJSON(geojson)
}
}

