% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wkt_reverse}
\alias{wkt_reverse}
\title{Reverses the points within a geometry.}
\usage{
wkt_reverse(x)
}
\arguments{
\item{x}{a character vector of WKT objects, represented as strings}
}
\value{
a string, same length as given
}
\description{
\code{wkt_reverse} reverses the points in any of
point, multipoint, linestring, multilinestring, polygon, or
multipolygon
}
\details{
segment, box, and ring types not supported
}
\examples{
wkt_reverse("POLYGON((42 -26,42 -13,52 -13,52 -26,42 -26))")
}
