% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as.json.R
\name{as.json}
\alias{as.json}
\alias{as.json.geojson}
\title{Convert geojson R list to JSON}
\usage{
as.json(x, pretty = TRUE, auto_unbox = TRUE, ...)

\method{as.json}{geojson}(x, pretty = TRUE, auto_unbox = TRUE, ...)
}
\arguments{
\item{x}{Output from \code{\link{wkt2geojson}}}

\item{pretty}{(logical) Adds indentation whitespace to JSON output. Can be TRUE/FALSE
or a number specifying the number of spaces to indent. See \code{\link[jsonlite]{prettify}}
Default: \code{TRUE} Having \code{TRUE} as default makes it easy to copy paste to a
text editor, etc.}

\item{auto_unbox}{(logical) Automatically unbox all atomic vectors of length 1.
Default: \code{TRUE}}

\item{...}{Further args passed on to \code{\link[jsonlite]{toJSON}}}
}
\description{
Convert geojson R list to JSON
}
\examples{
str <- "POLYGON ((100 0.1, 101.1 0.3, 101 0.5, 100 0.1),
   (103.2 0.2, 104.8 0.2, 100.8 0.8, 103.2 0.2))"
as.json(wkt2geojson(str))
as.json(wkt2geojson(str), FALSE)
}

