% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wkt_r.R
\name{dump_linestring}
\alias{dump_linestring}
\title{Convert GeoJSON-like LineString object to WKT.}
\usage{
dump_linestring(obj, fmt = 16)
}
\arguments{
\item{obj}{A GeoJSON-like object representing a Point, LineString, Polygon, MultiPolygon, etc.}

\item{fmt}{Format string which indicates the number of digits to display after the
decimal point when formatting coordinates.}
}
\description{
Convert GeoJSON-like LineString object to WKT.
}
\keyword{internal}

