% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wkt_r.R
\name{dump_point}
\alias{dump_point}
\title{Dump a GeoJSON-like Point object to WKT.}
\usage{
dump_point(obj, fmt)
}
\arguments{
\item{obj}{A GeoJSON-like `dict` representing a Point.}

\item{fmt}{Format string which indicates the number of digits to display after the
decimal point when formatting coordinates.}
}
\value{
WKT representation of the input GeoJSON Point ``obj``.
}
\description{
Dump a GeoJSON-like Point object to WKT.
}
\examples{
point <- list('type' = 'Point', 'coordinates' = c(116.4, 45.2, 11.1))
dump_point(point)
}

