% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wkt_r.R
\name{dump_polygon}
\alias{dump_polygon}
\title{Convert GeoJSON-like Polygon object to WKT.}
\usage{
dump_polygon(obj, fmt = 16)
}
\arguments{
\item{obj}{A GeoJSON-like `dict` representing a LineString}

\item{fmt}{Format string which indicates the number of digits to display after the
decimal point when formatting coordinates.}
}
\description{
Convert GeoJSON-like Polygon object to WKT.
}
\examples{
poly <- list(type = 'Polygon',
     coordinates=list(
       list(c(100.001, 0.001), c(101.12345, 0.001), c(101.001, 1.001), c(100.001, 0.001)),
       list(c(100.201, 0.201), c(100.801, 0.201), c(100.801, 0.801), c(100.201, 0.201))
))
dump_polygon(poly)
dump_polygon(poly, fmt=6)
}

