% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wkt_r.R
\name{geojson2wkt}
\alias{geojson2wkt}
\title{Convert GeoJSON-like objects to WKT.}
\usage{
geojson2wkt(obj, fmt = 16)
}
\arguments{
\item{obj}{A GeoJSON-like object representing a Point, LineString, Polygon, MultiPolygon, etc.}

\item{fmt}{Format string which indicates the number of digits to display after the
decimal point when formatting coordinates.}
}
\description{
Convert GeoJSON-like objects to WKT.
}
\examples{
# point
point <- list('type' = 'Point', 'coordinates' = c(116.4, 45.2))
geojson2wkt(point)

# linestring
st <- list(type = 'LineString',
           coordinates = list(c(0.0, 0.0, 10.0), c(2.0, 1.0, 20.0),
                             c(4.0, 2.0, 30.0), c(5.0, 4.0, 40.0)))
geojson2wkt(st)

# polygon
poly <- list(type = 'Polygon',
     coordinates=list(
       list(c(100.001, 0.001), c(101.12345, 0.001), c(101.001, 1.001), c(100.001, 0.001)),
       list(c(100.201, 0.201), c(100.801, 0.201), c(100.801, 0.801), c(100.201, 0.201))
))
geojson2wkt(poly)
geojson2wkt(poly, fmt=6)
}
\seealso{
\code{\link{wkt2geojson}}
}

