% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{load_point}
\alias{load_point}
\title{Convert WKT to GeoJSON-like POINT object.}
\usage{
load_point(str, fmt = 16, feature = TRUE)
}
\arguments{
\item{fmt}{Format string which indicates the number of digits to display after the
decimal point when formatting coordinates.}
}
\description{
Convert WKT to GeoJSON-like POINT object.
}
\keyword{internal}

