% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multipoint.R
\name{multipoint}
\alias{multipoint}
\title{Make WKT multipoint objects}
\usage{
multipoint(..., fmt = 16)
}
\arguments{
\item{...}{A GeoJSON-like object representing a Point, LineString, Polygon, MultiPolygon, etc.}

\item{fmt}{Format string which indicates the number of digits to display after the
decimal point when formatting coordinates. Max: 20}
}
\description{
Make WKT multipoint objects
}
\examples{
multipoint(c(100.000, 3.101), c(101.000, 2.100), c(3.140, 2.180))

# data.frame
library("maps")
data(us.cities)
df <- us.cities[1:25,c('long','lat')]
multipoint(df)
}

