% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{properties}
\alias{properties}
\title{Add properties to a geojson object}
\usage{
properties(x, style = NULL, popup = NULL)
}
\arguments{
\item{x}{Input}

\item{style}{List of color, fillColor, etc., or NULL}

\item{popup}{Popup text, or NULL}
}
\description{
Add properties to a geojson object
}
\examples{
str <- "POINT (-116.4000000000000057 45.2000000000000028)"
x <- wkt2geojson(str)
properties(x, style=list(color = "red"))
}
