% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_convert.R
\name{sp_convert}
\alias{sp_convert}
\alias{sf_convert}
\title{Convert spatial objects to WKT}
\usage{
sp_convert(x, group = TRUE)

sf_convert(x, group = TRUE)
}
\arguments{
\item{x}{for \code{sp_convert()}, a list of SP/SPDF objects (or a single object)
for \code{sf_convert()}, an sf, sfc, POLYGON, or MULTIPOLYGON sf object}

\item{group}{whether or not to group coordinates together in the case
that an object in \code{x} has multiple sets of coordinates. If \code{TRUE}
(the default), such objects will be returned as \code{MULTIPOLYGON}'s
\itemize{
\item if \code{FALSE}, as a vector of \code{POLYGON}'s
}}
}
\value{
either a character vector of WKT objects - one per sp object -
if \code{group} is \code{TRUE}, or a list of vectors if \code{group} is \code{FALSE}
}
\description{
\code{sp_convert} turns objects from the \code{sp} package
(SpatialPolygons, SpatialPolygonDataFrames) or the \code{sf} package
(sf, sfc, POLYGON, MULTIPOLYGON) - into WKT POLYGONs or MULTIPOLYGONs
}
\examples{
\dontrun{
library(sp)
library(sf)
s1 <- SpatialPolygons(list(Polygons(list(Polygon(cbind(c(2,4,4,1,2),c(2,3,5,4,2)))), "s1")))
sp_convert(s1)
x = st_as_sf(s1)
sf_convert(x)

library(sf)
one = Polygon(cbind(c(91,90,90,91), c(30,30,32,30)))
two = Polygon(cbind(c(94,92,92,94), c(40,40,42,40)))
spone = Polygons(list(one), "s1")
sptwo = Polygons(list(two), "s2")
z = SpatialPolygons(list(spone, sptwo), as.integer(1:2))
x = st_as_sf(z)
sp_convert(z)
# class: sf
sf_convert(x)
# class: sfc
sf_convert(x = x[[1]])
# class: polygon
sf_convert(x = unclass(x[[1]])[[1]])
sf_convert(x = unclass(x[[1]])[[2]])

library(silicate)
x <- sfzoo$multipolygon
class(x)
x_sp <- as(x, "Spatial")
st_as_text(x)
sp_convert(x_sp)
sf_convert(x)
}
}
\seealso{
\code{\link[=bounding_wkt]{bounding_wkt()}}, for turning bounding boxes within
\code{sp} objects into WKT objects.
}
