% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vis.R
\name{vis}
\alias{vis}
\title{Visualize well-known text area's on a map using leaflet and htmlwidgets packages}
\description{
Visualize well-known text area's on a map using leaflet and htmlwidgets packages
}
\examples{
\dontrun{
# Using htmlwidgets and leaflet packages
library("leaflet")
library("htmlwidgets")

## points
str <- "POINT (-116.4000000000000057 45.2000000000000028)"
geojson <- wkt2geojson(str) \%>\% properties(style=list(color = "red"))
leaflet() \%>\%
 addTiles() \%>\%
 addGeoJSON(geojson)

## polygons
str <- "POLYGON ((100 0.1, 101.1 0.3, 101 0.5, 100 0.1),
   (103.2 0.2, 104.8 0.2, 100.8 0.8, 103.2 0.2))"
geojson <- wkt2geojson(str) \%>\% properties(style=list(color = "red"))
cen <- centroid(geojson)
leaflet() \%>\%
 addTiles() \%>\%
 setView(lng = cen[2], lat = cen[1], zoom = 6) \%>\%
 addGeoJSON(geojson)

## linestring
geojson <- wkt2geojson("LINESTRING (0 -1, -2 -3, -4 5)", feature=TRUE) \%>\%
   properties(style=list(color = "red"))
leaflet() \%>\%
 addTiles() \%>\%
 setView(lng = -2, lat = 2, zoom = 6) \%>\%
 addGeoJSON(geojson)

st <- list(type = "Feature",
           geometry = list(
             type = 'LineString',
             coordinates = list(c(0.0, 0.0), c(2.0, 1.0),
                             c(4.0, 2.0), c(5.0, 4.0)))
)
st <- st \%>\% properties(style=list(color = "red"))
cen <- centroid(st)
leaflet() \%>\%
 addTiles() \%>\%
 setView(lng = cen[2], lat = cen[1], zoom = 6) \%>\%
 addGeoJSON(st)
}
}

