% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wellknown-package.R
\docType{package}
\name{wellknown-package}
\alias{wellknown}
\alias{wellknown-package}
\title{WKT to GeoJSON and vice versa}
\description{
WKT to GeoJSON and vice versa
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

