% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wkt2geojson.R
\name{wkt2geojson}
\alias{wkt2geojson}
\title{Convert WKT to GeoJSON-like objects.}
\usage{
wkt2geojson(str, fmt = 16, feature = TRUE)
}
\arguments{
\item{str}{A GeoJSON-like object representing a Point, LineString, Polygon, MultiPolygon, etc.}

\item{fmt}{Format string which indicates the number of digits to display after the
decimal point when formatting coordinates.}

\item{feature}{(logical) Make a feature geojson object. Default: TRUE}
}
\description{
Convert WKT to GeoJSON-like objects.
}
\examples{
# point
str <- "POINT (-116.4000000000000057 45.2000000000000028)"
wkt2geojson(str)
wkt2geojson(str, feature=FALSE)

# multipoint
str <- 'MULTIPOINT ((100.000 3.101), (101.000 2.100), (3.140 2.180))'
wkt2geojson(str)
wkt2geojson(str, feature=FALSE)

# polygon
str <- "POLYGON ((100 0.1, 101.1 0.3, 101 0.5, 100 0.1),
   (103.2 0.2, 104.8 0.2, 100.8 0.8, 103.2 0.2))"
wkt2geojson(str)
wkt2geojson(str, feature=FALSE)

# multipolygon
str <- "MULTIPOLYGON (((40 40, 20 45, 45 30, 40 40)),
   ((20 35, 45 20, 30 5, 10 10, 10 30, 20 35), (30 20, 20 25, 20 15, 30 20)))"
wkt2geojson(str)
wkt2geojson(str, feature=FALSE)

# linestring
str <- "LINESTRING (100.000 0.000, 101.000 1.000)"
wkt2geojson(str)
wkt2geojson(str, feature=FALSE)
wkt2geojson("LINESTRING (0 -1, -2 -3, -4 5)")
wkt2geojson("LINESTRING (0 1 2 3, 4 5 6 7)")
}
\seealso{
\code{\link{geojson2wkt}}
}

